refactor(react): extract shared ReactDocgenResolveError and RESOLVE_EXTENSIONS to internal/common
ReactDocgenResolveError and the docgen lookup logic were copy-pasted
identically between react-vite and react-webpack, with a warning comment
instructing developers to manually keep both files in sync.

Extract both to code/core/src/common/utils/docgen-resolver.ts and export
them via storybook/internal/common so both build tools import from one
place. The react renderer's docgenResolver.ts is updated to re-export
ReactDocgenResolveError from common (removing local duplication of the
class) while keeping its own defaultLookupModule that uses the broader
supportedExtensions list.

Closes #34332
