+++ code/core/src/common/utils/docgen-resolver.ts
@@ -1,19 +1,42 @@
 import { extname } from 'node:path';
 
-import { ReactDocgenResolveError, supportedExtensions } from 'storybook/internal/common';
-
 import resolve from 'resolve';
 
-export { ReactDocgenResolveError };
+export class ReactDocgenResolveError extends Error {
+  // the magic string that react-docgen uses to check if a module is ignored
+  readonly code = 'MODULE_NOT_FOUND';
+
+  constructor(filename: string) {
+    super(`'${filename}' was ignored by react-docgen.`);
+  }
+}
 
 /* The below code was copied from:
  * https://github.com/reactjs/react-docgen/blob/df2daa8b6f0af693ecc3c4dc49f2246f60552bcb/packages/react-docgen/src/importer/makeFsImporter.ts#L14-L63
  * because it wasn't exported from the react-docgen package.
  */
+
+// These extensions are sorted by priority
+// resolve() will check for files in the order these extensions are sorted
+export const RESOLVE_EXTENSIONS = [
+  '.js',
+  '.cts', // These were originally not in the code, I added them
+  '.mts', // These were originally not in the code, I added them
+  '.ctsx', // These were originally not in the code, I added them
+  '.mtsx', // These were originally not in the code, I added them
+  '.ts',
+  '.tsx',
+  '.mjs',
+  '.cjs',
+  '.mts',
+  '.cts',
+  '.jsx',
+];
+
 export function defaultLookupModule(filename: string, basedir: string): string {
   const resolveOptions = {
     basedir,
-    extensions: supportedExtensions,
+    extensions: RESOLVE_EXTENSIONS,
     // we do not need to check core modules as we cannot import them anyway
     includeCoreModules: false,
   };
